Refactor program structure and remove unused models
- Updated Program.cs to load default data for customers and veterinarians.
- Removed the Animal, Appointment, Customer, Pet, and Veterinarian models as they were no longer needed.
- Introduced DefaultData class to provide default customers and veterinarians.
- Implemented ConsoleHelper for consistent success and error messaging.
- Refactored CustomerService and VeterinaryService to improve functionality and maintainability.

+++ Models/Animal.cs
@@ -12,7 +12,7 @@ public abstract class Animal
     public string Type { get; set; } = string.Empty;
 
     // Edad inicial en 0
-    public int Age { get; set; } = 0;
+    public double Age { get; set; } = 0;
